Add notifications config option
Previously it wasnt possible to prevent loading the notifications plugin altogether. This update retains backwards compatibility with the _default_ settings for the module.

Breaking changes are in the form of removing the `posts-workflow` and `editorial-workflow` config options in favour of moving below a notifications options similar to the publication checklist config.

The update process is simply switching to the new configuration.

Fixes #25