use get_post_types to get an array of public post types
We can use the list of public post types as the default instead of an empty array. This way, the post types that are enabled are either those explicitly defined in the config, or all public post types by default.

adds a @uses parameter to the docblock to add a note that get_post_types must exist for the function to work properly. This shouldn't be an issue if the function is run normally in the admin but will break if loaded too early.