use a new post copy hook that's built into duplicate post
this hook runs after the new post is copied. 
we do a bunch of check to make sure we're only on a copied post and not a rewritten post, and then we make sure we haven't already updated the client ID. if all that passes, we update the client id and update the post, then store a meta value saying we did it.