copy pasta wholesale javascript functions from duplicate-post
many of these functions are not exported, so we can't import them. even if they are exported, their package doesn't include the original, unbuilt javascript files, and the compiled versions have version suffixes added to them, which would lock us into specific versions of the plugin. This allows us to not have to lock Duplicate Post at a specific version and update that version when each new release is made, we just need to watch for bugs/changes.